Updated without PHP hard requirements

+++ comp3550_setup.sh
@@ -1,37 +1,54 @@
 #!/bin/bash
 
+echo "Installing Base requirements: (Java, Sublime, Git)"
 
 sudo add-apt-repository -y ppa:webupd8team/sublime-text-3
 sudo add-apt-repository -y ppa:webupd8team/java
-sudo add-apt-repository -y ppa:paolorotolo/android-studio
-sudo add-apt-repository -y ppa:chris-lea/node.js 
-
 sudo apt-get update
+# Remove requirements for user confirmation of java liscense
+sudo echo oracle-java8-installer shared/accepted-oracle-license-v1-1 select true | /usr/bin/debconf-set-selections
+sudo apt-get install -y git oracle-java8-installer sublime-text-installer curl
 
-sudo apt-get install -y git oracle-java8-installer sublime-text-installer nodejs apache2 php5 php5-mysql mysql-server-5.6 mysql-client-5.6 phpmyadmin php5-mcrypt curl
 
-
-# Install dependencies for PHP development
-
-# Enable the Mcrypt for more robust encryption & hashing
-sudo php5enmod mcrypt
-sudo service apache2 restart
-
-#Install Yeoman Tool
-sudo npm install -g grunt-cli bower yo 
-
-# Install composer
-curl -sS https://getcomposer.org/installer | php
-sudo mv composer.phar /usr/local/bin/composer
-
-# Install Laravel
-read -p "The installation of Laravel takes some time. Will you like to install this now y/N " yn
+read -p "Install Nodejs y/N " yn
 case $yn in
-	[Yy]* ) sudo composer global require "laravel/installer=~1.1"
+	[Yy]* ) 
+		curl -sL https://deb.nodesource.com/setup_0.12 | sudo bash - && sudo apt-get install -y nodejs
 esac
 
+read -p "Install Command Line Interface (CLI) Tools y/N " yn
+case $yn in
+	[Yy]* ) 
+		echo "Installing Node. Node is a dependency for CLI tools"
+		curl -sL https://deb.nodesource.com/setup_0.12 | sudo bash - && sudo apt-get install -y nodejs
+		echo "Installing CLI Tools (Grunt, Bower, yo, Gulp)"
+		#Install Yeoman Tool
+		sudo npm install -g grunt-cli bower yo gulp
 
-mkdir ~/dev
+		echo "Testing packages"
+		yo --version
+		grunt --version
+		bower -v
+		gulp -v
+esac
+
+read -p "Install LAMP Stack y/N" yn
+case $yn in
+	[Yy]* ) 
+		sudo apt-get install -y apache2 php5 php5-mysql mysql-server-5.6 mysql-client-5.6 phpmyadmin php5-mcrypt
+		echo "Configuring Apache Installation"
+		sudo a2enmod rewrite
+		echo "Change the default.conf file to AllowOverride All"
+		sed -i '/AllowOverride None/c AllowOverride All' /etc/apache2/sites-available/default.confg
+		sudo php5enmod mcrypt
+		sudo service apache2 restart
+		# Install composer
+		echo "Installing Composer"
+		curl -sS https://getcomposer.org/installer | php
+		sudo mv composer.phar /usr/local/bin/composer
+		composer --version
+esac
+
